Created class initOgre and commented the file exporting line
I have extracted the class initOgre from the main and i have created the
files initOgre.h and initOgre.cpp.
I have also commented the line of code which creates the file mesh on
the computer since the path is not still general.
I have to find a way to choose the folder in which to download the file
dinamically.