Material import and resource.cfg file
Created the file resource.cfg. The program now, when it starts, use the
file resource.cfg to load all the resources contained in the folder
"media" (all the sub-folders have to be spefied).
In this version of the code, the mesh of the ram is loaded. Also the
materials relative to it are correctely loaded.