Fixed Ram Texture and Modified attachMesh()
The texture of the ram.mesh has been fixed and it's working now. The
texture of char_ram_nor.png is not used anymore.
Now the function attachMesh() use the string in argument to give the
name at the node.
Two objects are now attached to the planet, so there's a new mesh with
the texture still not working (not important issue).