2D-grid movement support
Adds nextPosition(), which gives new 3D-position while taking input as 4
different directions

Caveats:
Relies heavily on HeightMap, so edges of adjacent grids have same 3D-
position

Size of the grids are same as the HeightMaps to construct geometry of
the sphere.

Movements on z+ and z- -grid might not go as expected, for example,
moving in continuously in y+ and going from y- to z+-grid is going to
pingpong between the two grids.